Required coefficents to be integral.

+++ src/sci/chem/reaction.rs
@@ -7,22 +7,22 @@ use attr_mac::json;
 #[json]
 pub struct Reaction {
     /// List of reactant species buy id, and their stoichiometric coefficient.
-    reactants: Vec<(Name, f64)>,
+    reactants: Vec<(Name, i32)>,
     /// List of product species buy id, and their stoichiometric coefficient.
-    products: Vec<(Name, f64)>,
+    products: Vec<(Name, i32)>,
     /// Rate.
     rate: Rate,
 }
 
 impl Reaction {
-    access!(reactants, Vec<(Name, f64)>);
-    access!(products, Vec<(Name, f64)>);
+    access!(reactants, Vec<(Name, i32)>);
+    access!(products, Vec<(Name, i32)>);
     access!(rate, Rate);
 
     /// Construct a new instance.
     #[inline]
     #[must_use]
-    pub fn new(reactants: Vec<(Name, f64)>, products: Vec<(Name, f64)>, rate: Rate) -> Self {
+    pub fn new(reactants: Vec<(Name, i32)>, products: Vec<(Name, i32)>, rate: Rate) -> Self {
         Self {
             reactants,
             products,
